requirements.txt
camelot-py
pytest
PyPDF2
python-docx
python-pptx
PyMuPDF
pandas
opencv-python
ghostscript
fpdf
lorem
msoffcrypto-tool
python-dotenv